+++ CHANGELOG.md
@@ -7,10 +7,15 @@ v5.7.0
 - Fix memory-leak with LowEnergyBlock preconditioner for time-updated matrices (!1627)
 - Fix Fourier expansion integration weights are related test (!1803)
 - Separate MeshGraph input/output functions into a new class (!1778)
+- Added checkpoint file writing start time in the fieldconvert filter (!1789)
+- Fix fieldconvert filter incorrect boundary values (!1789)
+- Fix numerical precision issues with filters OutputStartTime (!1789)
 - Fix AdaptiveSFD for MPI (!1821)
+- Fix IterativeStaticCond when using absolute tolerance (!1850)
+- Fix deadlock by scotch with multi-threading support (!1853)
 
 **IncNavierStokesSolver**
-- Fix initial and boundary conditions in the moving reference frame (!1692)
+- Fix initial and boundary conditions in the moving reference frame (!1692, !1820)
 - Fix memory-leak for the Mixed_CG_Discontinuous projection when initializing the traceMep (!1806)
 - Add synthetic turbulence generation for the incompressible solver (!1664) 
 
@@ -19,8 +24,10 @@ v5.7.0
 
 **NekMesh**
 - Added revolve module (!1825)
+
 **FieldConvert**
 - Add vortexinducedvelocity module to compute the vortex-induced velocity (!1824)
+- Add a module to transform coordinates and vectors for the moving reference frame method (!1830)
 
 **Miscellaneous**
 - Use std::stod instead of boost::lexical_cast<NekDouble> (!1819)